chore: add strict compiler/checkstyle guardrails (Xlint+Werror, Error Prone, Checkstyle)
Adds -Xlint:all -Werror and Google Error Prone to maven-compiler-plugin,
and maven-checkstyle-plugin 3.3.1 (bumped from the unused 2.17 pluginManagement
entry) wired into ci-harness.sh's linter gate alongside spotbugs:check.

checkstyle.xml required several compatibility fixes for the modern Checkstyle
core bundled with 3.3.1 vs. the original ~2010-era config: LineLength moved
from a TreeWalker to a Checker child, SuppressionCommentFilter/
SuppressWithNearbyCommentFilter moved the other way, JavadocMethod's
allowMissing*/logLoadErrors properties were removed (that check now only
validates tags on javadoc that already exists, it can no longer be told
javadoc is optional), and resource-file scanning was disabled since it
was flagging translation .properties files for line length. JavadocMethod
and JavadocVariable were dropped rather than left enforcing full javadoc-tag
completeness project-wide, which was never the original config's intent and
is a separate, much larger task than this guardrail's scope. Added a
RegexpSinglelineJava rule forbidding `var x = someCall();` (method-call RHS
with a non-obvious type), per the task's requirement.

Error Prone needs JDK17 module access (-J--add-exports/--add-opens for
several jdk.compiler internal packages) to run at all; maven.compiler.fork
was already true from prior work, which those -J flags depend on.

Fixed the first batch of -Werror findings this surfaced (40 Xlint warnings,
then Error Prone findings revealed after those: MissingOverride across 9
files' interface-implementing methods, ReferenceEquality, JdkObsolete,
JavaUtilDate, an intentional Comparable<ReadablePartial> cross-type design
in AbstractPartial, and a few others) to validate the new gates actually
work end to end, following the same fix-vs-suppress-with-justification
judgment used for the pre-existing SpotBugs baseline. The full remaining
scope surfaced only after removing the default javac -Xmaxwarns 100 cap:
412 total -Werror findings across ~20 Error Prone categories, plus 37
checkstyle violations after the above config fixes — both far larger than
what fit in this pass, remediation continues in a follow-up commit.

+++ src/main/java/com/legacy/system/datetime/format/DateTimeFormatterBuilder.java
@@ -2061,7 +2061,7 @@ public class DateTimeFormatterBuilder {
       if (iMaxFields == 1) {
         return;
       }
-      displayOffset -= hours * (int) DateTimeConstants.MILLIS_PER_HOUR;
+      displayOffset -= hours * DateTimeConstants.MILLIS_PER_HOUR;
       if (displayOffset == 0 && iMinFields <= 1) {
         return;
       }
